Add files via upload
 All methods within BankAccount Class must be implemented.

Create a total of 1,000 banking accounts in the main class Q4. Randomly allocate an initial balance between 0 and 50,000 for each bank account. Also, assign ownerName sequentially as "owner_1, owner_2, ..., owner_1000".

Please add $10,000 to everyone's banking account.

Please print out the owner's name and the balance of the person with the largest balance out of 1,000 bank accounts.